refactor: 추상화 - view controller, DOMelements 파일 위치 - intersection observer 함수 분리
